+++ services/transaction.go
@@ -22,9 +22,9 @@ func NewTxnOps(ctx context.Context, app *config.Application) *TxnOps {
 	}
 }
 
-func (r *TxnOps) GenerateTransaction(groupObj *ent.Group, sourceUser *ent.User, destUser *ent.User, amount int, note string) (*ent.Transaction, error) {
-	existingLentTxn := 0
-	existingOwedTxn := 0
+func (r *TxnOps) GenerateTransaction(groupObj *ent.Group, sourceUser *ent.User, destUser *ent.User, amount float64, note string) (*ent.Transaction, error) {
+	existingLentTxn := 0.0
+	existingOwedTxn := 0.0
 	var err error
 	if temp := r.app.EntClient.Transaction.Query().
 		Where(
@@ -41,7 +41,7 @@ func (r *TxnOps) GenerateTransaction(groupObj *ent.Group, sourceUser *ent.User, 
 					transaction.HasDestinationWith(user2.IDEQ(destUser.ID)),
 				),
 				transaction.HasBelongsToWith(group.IDEQ(groupObj.ID)),
-			).Aggregate(ent.Sum(transaction.FieldAmount)).Int(r.ctx)
+			).Aggregate(ent.Sum(transaction.FieldAmount)).Float64(r.ctx)
 		if err != nil {
 			log.Error(err)
 			return nil, err
@@ -62,7 +62,7 @@ func (r *TxnOps) GenerateTransaction(groupObj *ent.Group, sourceUser *ent.User, 
 					transaction.HasSourceWith(user2.IDEQ(destUser.ID)),
 				),
 				transaction.HasBelongsToWith(group.IDEQ(groupObj.ID)),
-			).Aggregate(ent.Sum(transaction.FieldAmount)).Int(r.ctx)
+			).Aggregate(ent.Sum(transaction.FieldAmount)).Float64(r.ctx)
 		if err != nil {
 			return nil, err
 		}
